requirements.txt
pyyaml
bottle
